+++ lib/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "typst-ansi-hl-lib"
+version = { workspace = true }
+edition = { workspace = true }
+authors = { workspace = true }
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+bat = { workspace = true }
+termcolor = { workspace = true }
+thiserror = { workspace = true }
+typst-syntax = { workspace = true }
